requirements_test.txt
pytest==3.2.1
flake8==3.4.1
pytest-cov==2.5.1
mock==2.0.0
flask-testing==0.6.2
